racing-api-chatbot/requirements.txt
openai
langgraph
python-dotenv
requests
backoff>=2.2.1
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
langchain
langchain-openai
langchain-core
python-decouple
psycopg2-binary
sqlalchemy>=2.0.24
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0